refactor GET meal_foods
