+++ lib/controllers/meals_controller.js
@@ -1,8 +1,6 @@
 const Meal = require('../models/meal')
 const Goal = require('../models/goal')
 
-const pry = require('pryjs')
-
 const index = (request, response) => {
   Meal.all()
     .then((meals_data) => {
