+++ lib/models/meal.js
@@ -18,6 +18,13 @@ const find_by_id = (id) => database('meals')
 const create = (name, date, calorie_goal) => database('meals')
   .insert({name: name, date: date, calorie_goal: calorie_goal}, 'id')
 
+const foods_by_meal_id = (id) => database('meals')
+  .leftJoin('meal_foods', 'meal_foods.meal_id', '=', 'meals.id')
+  .leftJoin('foods', 'meal_foods.food_id', '=', 'foods.id')
+  .where('meal_foods.meal_id', id)
+  .select( 'meals.id AS meal_id', 'meals.name AS meal_name', 'meals.date AS meal_date', 'meals.calorie_goal AS meal_goal', 'foods.id AS food_id', 'foods.name AS food_name', 'foods.calories AS food_calories')
+  
+
 module.exports = {
-  all, create, find_by_id, find_by_meal_and_date
+  all, create, find_by_id, find_by_meal_and_date, foods_by_meal_id
 }
