+++ lib/models/meal_food.js
@@ -1,0 +1,9 @@
+const environment = process.env.NODE_ENV || 'development'
+const configuration = require('../../knexfile')[environment];
+const database = require('knex')(configuration);
+
+const all = () => database('meal_foods')
+
+module.exports = {
+  all
+}
