+++ lib/routes/api/v1/meal_foods.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const MealFoodsController = require('../../../controllers/meal_foods_controller');
+
+router.get('/:meal_id/foods', MealFoodsController.index);
+
+module.exports = router
